Remove unused definitions and discard `token` submodule
